Concatenate TS so as not to add a module system
