--- src/wasm-js/index.ts
@@ -1,3 +1,0 @@
-export { readValue } from './read';
-export { writeValue } from './write';
-export { init } from './shared';
